drcu_fd/drcu_fd_pc_test_scripts/requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
cachetools==5.3.0
cffi==1.15.1
cryptography==40.0.2
decorator==5.1.1
fabric==3.0.0
future==0.18.3
ifaddr==0.2.0
invoke==2.0.0
iso8601==1.1.0
paramiko==3.1.0
pefile==2023.2.7
py2exe==0.13.0.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
pythoncrc==1.21
pytz==2023.3
PyYAML==6.0
scp==0.14.5
tenma-serial==1.1.0
zeroconf==0.62.0
